Move and test file objects.
