+++ ocean_lib/aquarius/test/test_aquarius.py
@@ -4,12 +4,11 @@
 #
 import pytest
 
-from ocean_lib.agreements.file_objects import FilesTypeFactory
 from ocean_lib.aquarius.aquarius import Aquarius
 from ocean_lib.assets.asset import Asset
 from ocean_lib.assets.asset_resolver import resolve_asset
-from ocean_lib.data_provider.data_service_provider import DataServiceProvider
 from ocean_lib.models.models_structures import CreateErc20Data
+from ocean_lib.structures.file_objects import FilesTypeFactory
 from ocean_lib.web3_internal.constants import ZERO_ADDRESS
 from ocean_lib.web3_internal.currency import to_wei
 from tests.resources.helper_functions import get_address_of_type
