+++ ocean_lib/data_provider/test/test_data_service_provider.py
@@ -10,7 +10,6 @@ import pytest
 from requests.exceptions import InvalidURL
 from requests.models import Response
 
-from ocean_lib.agreements.file_objects import FilesTypeFactory
 from ocean_lib.agreements.service_types import ServiceTypes
 from ocean_lib.data_provider.data_service_provider import DataServiceProvider as DataSP
 from ocean_lib.data_provider.data_service_provider import urljoin
@@ -18,6 +17,7 @@ from ocean_lib.exceptions import DataProviderException
 from ocean_lib.http_requests.requests_session import get_requests_session
 from ocean_lib.models.compute_input import ComputeInput
 from ocean_lib.services.service import Service
+from ocean_lib.structures.file_objects import FilesTypeFactory
 from tests.resources.ddo_helpers import create_basics
 from tests.resources.helper_functions import (
     deploy_erc721_erc20,
