+++ ocean_lib/ocean/test/test_ocean_assets.py
@@ -11,7 +11,6 @@ from unittest.mock import patch
 import eth_keys
 import pytest
 
-from ocean_lib.agreements.file_objects import FilesTypeFactory
 from ocean_lib.agreements.service_types import ServiceTypes
 from ocean_lib.assets.asset import Asset
 from ocean_lib.data_provider.data_service_provider import DataServiceProvider
@@ -19,11 +18,12 @@ from ocean_lib.exceptions import AquariusError, ContractNotFound, InsufficientBa
 from ocean_lib.models.erc721_factory import ERC721FactoryContract
 from ocean_lib.models.erc721_nft import ERC721NFT
 from ocean_lib.models.models_structures import (
+    ConsumeFees,
     CreateErc20Data,
     CreateERC721Data,
-    ConsumeFees,
 )
 from ocean_lib.services.service import Service
+from ocean_lib.structures.file_objects import FilesTypeFactory
 from ocean_lib.web3_internal.constants import ZERO_ADDRESS
 from ocean_lib.web3_internal.currency import to_wei
 from ocean_lib.web3_internal.wallet import Wallet
